Fix gap-follow: costmap dtype, OOB handling, no-gap fallback
- costmap_data stored as int16 (int8 wraps values >127)
- Out-of-bounds cells treated as free (unexplored), not blocked
- Unknown cells (-1) treated as free
- No-gap fallback: rotate in place at 0.5 rad/s (was barely moving)
- Higher minimum speed (0.2 * max instead of 0.15)
- Track max_free correctly per ray

+++ mini_r1_v1_application/scripts/maze_navigator_node.py
@@ -97,7 +97,7 @@ class MazeNavigatorNode(Node):
 
     # ── Callbacks ────────────────────────────────────────────────────────
     def costmap_cb(self, msg: OccupancyGrid):
-        self.ss.costmap_data = np.array(msg.data, dtype=np.int8)
+        self.ss.costmap_data = np.array(msg.data, dtype=np.int16)  # int8 wraps >127
         self.ss.costmap_width = msg.info.width
         self.ss.costmap_height = msg.info.height
         self.ss.costmap_resolution = msg.info.resolution
